requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
crispy-bootstrap5==0.7
dj-database-url==3.0.1
Django==4.2.7
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
